move ref.go to its own subpackage
