+++ pkg/api/BUILD
@@ -16,16 +16,16 @@ go_library(
         "doc.go",
         "field_constants.go",
         "json.go",
-        "methods.go",
-        "ref.go",
+        "objectreference.go",
         "register.go",
         "resource_helpers.go",
+        "taint.go",
+        "toleration.go",
         "types.go",
         "zz_generated.deepcopy.go",
     ],
     tags = ["automanaged"],
     deps = [
-        "//vendor:k8s.io/apimachinery/pkg/api/meta",
         "//vendor:k8s.io/apimachinery/pkg/api/resource",
         "//vendor:k8s.io/apimachinery/pkg/apimachinery/announced",
         "//vendor:k8s.io/apimachinery/pkg/apimachinery/registered",
@@ -44,17 +44,14 @@ go_library(
 go_test(
     name = "go_default_test",
     srcs = [
-        "methods_test.go",
-        "ref_test.go",
         "resource_helpers_test.go",
+        "taint_test.go",
     ],
     library = ":go_default_library",
     tags = ["automanaged"],
     deps = [
         "//vendor:k8s.io/apimachinery/pkg/api/resource",
         "//vendor:k8s.io/apimachinery/pkg/apis/meta/v1",
-        "//vendor:k8s.io/apimachinery/pkg/runtime",
-        "//vendor:k8s.io/apimachinery/pkg/runtime/schema",
     ],
 )
 
@@ -120,6 +117,7 @@ filegroup(
         "//pkg/api/install:all-srcs",
         "//pkg/api/meta:all-srcs",
         "//pkg/api/pod:all-srcs",
+        "//pkg/api/ref:all-srcs",
         "//pkg/api/resource:all-srcs",
         "//pkg/api/service:all-srcs",
         "//pkg/api/testapi:all-srcs",
