+++ pkg/client/clientset_generated/internalclientset/typed/core/internalversion/BUILD
@@ -38,6 +38,7 @@ go_library(
     tags = ["automanaged"],
     deps = [
         "//pkg/api:go_default_library",
+        "//pkg/api/ref:go_default_library",
         "//pkg/api/v1:go_default_library",
         "//pkg/client/clientset_generated/internalclientset/scheme:go_default_library",
         "//vendor:k8s.io/apimachinery/pkg/apis/meta/v1",
